Odstraněn přebytečný kod z elementů

+++ server/src/routes/posts/deletePost.js
@@ -1,28 +1,58 @@
+// const route = require("express").Router();
+// const modelPost = require("../../models/post");
+
+// route.delete("/delete-post", async (req, res) => {
+//     try{
+//         const { _id, autorID } = req.body;
+//         // To:DO - findOne?
+//         // TO:DO - lze udelat i pro comment?
+//         const post = await modelPost.find({ _id, autorID }); 
+        
+//         if(post){
+//             // await post.delete();// Smazání příspěvku
+//             await modelPost.deleteOne({ _id, autorID });
+
+//             res.status(200).json({
+//                 success:true,
+//                 msg:"Přispěvěk Odstraněn",
+//             })        
+//         }else{
+//             res.status(400).json({ msg: "Nejste Autorem Příšpeveku. Přispěvěk nebude odstraněn."})
+//         }
+//     } catch(err) {
+//         res.status(500).send({ msg: `Chyba: ${err}` })
+//     }
+// })
+
+// module.exports = route;
+
 const route = require("express").Router();
 const modelPost = require("../../models/post");
 
 route.delete("/delete-post", async (req, res) => {
-    try{
+    try {
         const { _id, autorID } = req.body;
-        // To:DO - findOne?
-        // TO:DO - lze udelat i pro comment?
-        const post = await modelPost.find({ _id, autorID }); 
+        const post = await modelPost.findById(_id); 
         
-        if(post){
-            // await post.delete();// Smazání příspěvku
-            await modelPost.deleteOne({ _id, autorID }); // Použijte deleteOne pro smazání
-
+        if (post && post.autorID.toString() === autorID) {
+            await modelPost.deleteOne({ _id, autorID });
 
             res.status(200).json({
-                msg:"Přispevěk Odstraněn",
-                success:true
-            })        
-        }else{
-            res.status(400).json({ msg: "Nejste Autorem Příšpeveku. Přispěvěk NEondstraněn."})
+                success: true,
+                msg: "Příspěvek byl úspěšně smazán",
+            });
+        } else {
+            res.status(400).json({
+                success: false,
+                msg: "Nejste autorem příspěvku nebo byl příspěvek již smazán.",
+            });
         }
-    } catch(err) {
-        res.status(500).send({ msg: `Chyba: ${err}` })
+    } catch (err) {
+        res.status(500).json({
+            success: false,
+            msg: `Chyba: ${err.message}`,
+        });
     }
-})
+});
 
 module.exports = route;
